content: move Go vanity for cuelabs.dev/go/oci to GerritHub
GitHub is now simply a git mirror.
Note that we still redirect the root of the repo to GitHub,
since pkg.go.dev only serves content for modules,
and we don't have any module at the root of the repo - just a workspace.

Generated as of f8d5c1d5c2.